+++ src/constant.h
@@ -63,14 +63,9 @@ inline unsigned int toIndex(BuiltinVarOffset offset) { return static_cast<unsign
 // =====  builtin variable  =====
 
 // shell builtin
-constexpr const char *VAR_EIDT_HOOK = "EDIT_HOOK";
 constexpr const char *VAR_LINE_EDIT = "LINE_EDIT";
 constexpr const char *VAR_CMD_FALLBACK = "CMD_FALLBACK";
 constexpr const char *VAR_DEF_SIGINT = "_DEF_SIGINT";
-
-constexpr const char *VAR_HISTFILE = "HISTFILE";
-constexpr const char *VAR_HISTSIZE = "HISTSIZE";
-constexpr const char *VAR_HISTFILESIZE = "HISTFILESIZE";
 
 constexpr const char *VAR_SIG_DFL = "SIG_DFL";
 constexpr const char *VAR_SIG_IGN = "SIG_IGN";